+++ slash/reporting/reporter_interface.py
@@ -1,5 +1,11 @@
 class ReporterInterface(object):
 
+    def notify_before_console_output(self):
+        pass
+
+    def notify_after_console_output(self):
+        pass
+
     def report_session_start(self, session):
         pass